requirements.txt
#python==3.10.11
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.17
google-generativeai==0.8.3
python-dotenv==1.0.1
PyPDF2==3.0.